+++ src/app.c
@@ -30,8 +30,14 @@ App *initialize_app(int argc, char **argv)
                      G_CALLBACK(authentication_complete_cb), app);
     g_signal_connect(GTK_ENTRY(APP_PASSWORD_INPUT(app)), "activate",
                      G_CALLBACK(handle_password), app);
-    g_signal_connect(GTK_WIDGET(APP_BACKGROUND_WINDOW(app)), "key-press-event",
-                     G_CALLBACK(handle_tab_key), app);
+    // This was added to fix a bug where the background window would be focused
+    // instead of the main window, preventing users from entering their password.
+    // It's undocument & probably not necessary any more. Investigate & remove.
+    for (int m = 0; m < APP_MONITOR_COUNT(app); m++) {
+        g_signal_connect(GTK_WIDGET(APP_BACKGROUND_WINDOWS(app)[m]),
+                         "key-press-event",
+                         G_CALLBACK(handle_tab_key), app);
+    }
     g_signal_connect(GTK_WIDGET(APP_MAIN_WINDOW(app)), "key-press-event",
                      G_CALLBACK(handle_power_management_keys), app->config);