+++ src/app.h
@@ -21,7 +21,8 @@ void destroy_app(App *app);
 #define APP_LOGIN_USER(app)             (app)->config->login_user
 
 /* UI Member Accessors */
-#define APP_BACKGROUND_WINDOW(app)      (app)->ui->background_window
+#define APP_BACKGROUND_WINDOWS(app)     (app)->ui->background_windows
+#define APP_MONITOR_COUNT(app)          (app)->ui->monitor_count
 #define APP_MAIN_WINDOW(app)            (app)->ui->main_window
 #define APP_PASSWORD_INPUT(app)         (app)->ui->password_input
 #define APP_FEEDBACK_LABEL(app)         (app)->ui->feedback_label