+++ src/ui.c
@@ -14,9 +14,9 @@
 
 
 static UI *new_ui(void);
-static void setup_background_window(UI *ui);
-static void set_window_to_screen_size(GtkWindow *window);
-static void get_screen_dimensions(GdkRectangle *geometry);
+static void setup_background_windows(UI *ui);
+static GtkWindow *new_background_window(GdkMonitor *monitor);
+static void set_window_to_monitor_size(GdkMonitor *monitor, GtkWindow *window);
 static void setup_main_window(Config *config, UI *ui);
 static void place_main_window(GtkWidget *main_window, gpointer user_data);
 static void create_and_attach_layout_container(UI *ui);
@@ -30,7 +30,7 @@ UI *initialize_ui(Config *config)
 {
     UI *ui = new_ui();
 
-    setup_background_window(ui);
+    setup_background_windows(ui);
     setup_main_window(config, ui);
     create_and_attach_layout_container(ui);
     create_and_attach_password_field(config, ui);
@@ -48,7 +48,8 @@ static UI *new_ui(void)
     if (ui == NULL) {
         g_error("Could not allocate memory for UI");
     }
-    ui->background_window = NULL;
+    ui->background_windows = NULL;
+    ui->monitor_count = 0;
     ui->main_window = NULL;
     ui->layout_container = NULL;
     ui->password_label = NULL;
@@ -59,8 +60,32 @@ static UI *new_ui(void)
 }
 
 
-/* Create & Configure the Background Window */
-static void setup_background_window(UI *ui)
+/* Create a Background Window for Every Monitor */
+static void setup_background_windows(UI *ui)
+{
+    GdkDisplay *display = gdk_display_get_default();
+    ui->monitor_count = gdk_display_get_n_monitors(display);
+    ui->background_windows = malloc((uint) ui->monitor_count * sizeof (GtkWindow *));
+    for (int m = 0; m < ui->monitor_count; m++) {
+        GdkMonitor *monitor = gdk_display_get_monitor(display, m);
+        if (monitor == NULL) {
+            break;
+        }
+
+        GtkWindow *background_window = new_background_window(monitor);
+        ui->background_windows[m] = background_window;
+
+        if (gdk_monitor_is_primary(monitor)) {
+            GtkStyleContext *style_context =
+                gtk_widget_get_style_context(GTK_WIDGET(background_window));
+            gtk_style_context_add_class(style_context, "with-image");
+        }
+    }
+}
+
+
+/* Create & Configure a Background Window for a Monitor */
+static GtkWindow *new_background_window(GdkMonitor *monitor)
 {
     GtkWindow *background_window = GTK_WINDOW(gtk_window_new(
         GTK_WINDOW_TOPLEVEL));
@@ -68,51 +93,29 @@ static void setup_background_window(UI *ui)
     gtk_window_set_keep_below(background_window, TRUE);
     gtk_widget_set_name(GTK_WIDGET(background_window), "background");
 
-    // Set Window Size to Screen Size
-    set_window_to_screen_size(background_window);
+    // Set Window Size to Monitor Size
+    set_window_to_monitor_size(monitor, background_window);
 
-
+    // TODO: is this needed?
     g_signal_connect(background_window, "destroy", G_CALLBACK(gtk_main_quit),
                      NULL);
 
-    ui->background_window = background_window;
+    return background_window;
 }
 
 
 /* Set the Window's Minimum Size to the Default Screen's Size */
-static void set_window_to_screen_size(GtkWindow *window)
+static void set_window_to_monitor_size(GdkMonitor *monitor, GtkWindow *window)
 {
     GdkRectangle geometry;
-    get_screen_dimensions(&geometry);
+    gdk_monitor_get_geometry(monitor, &geometry);
     gtk_widget_set_size_request(
         GTK_WIDGET(window),
         geometry.width,
         geometry.height
     );
-    gtk_window_move(window, 0, 0);
+    gtk_window_move(window, geometry.x, geometry.y);
     gtk_window_set_resizable(window, FALSE);
-}
-
-
-/* Naively Estimate the Total Screen Size by Summing Every Monitor Dimension */
-static void get_screen_dimensions(GdkRectangle *geometry)
-{
-    int width = 0;
-    int height = 0;
-    GdkDisplay *display = gdk_display_get_default();
-    int monitor_count = gdk_display_get_n_monitors(display);
-    for (int m = 0; m < monitor_count; m++) {
-        GdkMonitor *monitor = gdk_display_get_monitor(display, m);
-        if (monitor == NULL) {
-            break;
-        }
-        GdkRectangle monitor_geometry;
-        gdk_monitor_get_geometry(monitor, &monitor_geometry);
-        width += monitor_geometry.width;
-        height += monitor_geometry.height;
-    }
-    geometry->width = width;
-    geometry->height = height;
 }
 
 
@@ -237,7 +240,10 @@ static void attach_config_colors_to_screen(Config *config)
             "color: %s;\n"
         "}\n"
         "#background {\n"
-            "background-image: image(url(%s), %s);\n"
+            "background-color: %s;\n"
+        "}\n"
+        "#background.with-image {\n"
+            "background-image: image(url(%s));\n"
             "background-repeat: no-repeat;\n"
             "background-position: center;\n"
         "}\n"
@@ -261,8 +267,8 @@ static void attach_config_colors_to_screen(Config *config)
         , config->font
         , gdk_rgba_to_string(config->text_color)
         , gdk_rgba_to_string(config->error_color)
-        , config->background_image
         , gdk_rgba_to_string(config->background_color)
+        , config->background_image
         , config->border_width
         , gdk_rgba_to_string(config->border_color)
         , gdk_rgba_to_string(config->window_color)